+++ devenv.nix
@@ -112,6 +112,7 @@ in
       packages = packagesWithTests;
       vitestBin = "packages/@overeng/utils/node_modules/.bin/vitest";
       vitestConfig = "packages/@overeng/utils/vitest.config.ts";
+      vitestInstallTask = "pnpm:install:utils";
       extraTests = [ "nix:test" ];
     })
     (taskModules.storybook {